requirements.txt
asgiref==3.3.1
azure-common==1.1.26
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.4
dj-database-url==0.5.0
Django==3.1.6
django-appconf==1.0.4
django-crispy-forms==1.11.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.11.1
django-utils-six==2.0
idna==2.10
install==1.3.4
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
gunicorn==19.9.0
django-ckeditor==6.0.0
